add i18n support
i18next was chosen as the translation framework because it is very
easy to integrate, and has nice batteries included (e.g. an in-app
translation editor to translate right inside the app, easy integration
with an online collaboration tool, etc.).

In this first draft, translations are kept in the web app in hard
coded files, but it might move to the backend and ne served over http
in the future.