wallet: fix init wallet
It was initialized after login, which does not work if the device is
not initialized, resulting in a panic.

Now it is initialized after the device status changes to seeded, which
is after login if the wallet is seeded, or after wallet creation/restoration.

It also fixes the issue where the wallet was not reinitialized when
restored from backup.