knot: use proper types, not strings, for events.
First step in removing some stringified types. device.Status() strings
follow.

At the same time, the knot is more directly forwarding the events of
the components. Instead of the websocket feeding the data directly, it
now only notifies of a change, so the client needs to call the
relevant endpoint to learn more. This is to avoid duplication, and the
extra roundtrip does not matter for performance as everything is local.