device/bitbox: automatically lock bootloader
If the device bootloader is unlocked, it is automatically locked after
login, which removes the LED lightning up and giving the user the
option to enter the bootloader.

If the bootloader is needed, it can be unlocked after login, und
entered with a replug before a new login.

This reduces confusion for the user and UI overhead in locking the
bootloader after using it.