authorize-client: restrict communication with the app backend
Only an authorized client may talk to the app's backend. Access to the
backend is restricted by requiring an authorization token to be sent
along with the API call. The authorization token must have been issued
previously by the backend and injected into the frontend.

The token is sent via HTTP header (Authorize).
The backend injects an authorization token into the QT app.
All API and asset requests, as well as the establishing of the websocket, are
protected by this token.
When in dev mode (= in the browser), the token is not required.

Until the client of the websocket is authorized, all messages are buffered
in the backend and send to the client once it is authorized.