wallet: persist transaction index in database
All downloaded address histories, transactions and extracted
inputs/outputs are cached in a local database. With this, after the
wallet is restarted, the histories and transactions touching the
wallet do not have to be downloaded and indexed again.

The DB used is the key/value database bbolt. The database is accessed
via an interface, so the db backend can easily be swapped out if
needed.

Database transactions are used. The software locks are kept to
serialize the transactions, so there can be no database transaction
commit errors (pessimistic locking over optimistic locking for now,
can switch later if needed).

The database is stored in a fixed location; it will be moved to a
application folder along with the certificates and logging.