wallet: decouple wallet addresses from transactions indexing
Before, the wallet would sneak in additional info via
addresses.Address to the transactions package, and get it back out
again for signing transactions via a type coercion. This is one step
into decoupling the two.

With this, the transactions package can independently serialize/store
the indexed data in a database.