frontend: fix firmware upgrade (and more)
When switching between registered/unregistered, the deviceStatus was
not reset, resulting in a temporary rerending of the wrong screen:

When going from a logged in device to the bootloader, the deviceStatus
was still SEEDED, and <Seed/> was rendered, making invalid api calls
targeting the device, not the bootloader.

Vice versa, going from the bootloader to the login screen had the
lingering BOOTLOADER device status, rendering the bootloader before
updating the status, resulting also in invalid api calls.

This resulted in wrong commands being sent to the device (firmware
cmds in bootloader and vice versa), which hangs the device. To prevent
this, all device commands now error if called in the wrong context.