jsonrpc: handle response errors
Before, all errors were fed into the failureCallback. If the error was
in the response to a specific request, then the response callback was
never called, which is a problem if the client is waiting on a
response.

Specifically, BroadcastTransaction() waited on a response, and if the
fee minrelay rule was not met, it would wait until the rpc response
timeout with a non descriptive message.