account: fixes error when BitBox is unplugged
A previous commit introduced a bug that triggered an /Init call on
the account if the BitBox was unplugged. To fix it, the account
know also has the status "keystoreAvailable". Only if the account
is not initialized and the keystore (BitBox or software) is available,
the /Init call is performed.

Overview of the account stati:
- offline / online (= backend can connect to the blockchain)
- uninitialized / initial synchronization done (= account data was fully synchronized with the blockchain)
- keystore unavailable / keystore available (= BitBox is plugged in)