docker: Check for godbb image and give hint if it is missing in dockerdev.sh
A user that runs dockerdev.sh (or 'make dockerdev') without this change gets an
error like `Unable to find image 'godbb:latest' locally`.

With this change, the user instead gets a friendly hint to run the 'make dockerinit'
command (which is mentioned in the docs, but not all users read them it turns
out.. :)).