util/config: check the app config dir only once
There's no reason why AppDir should do filesystem stat calls
every time it is called. It is enough to do it once at
the startup and reuse the value for the entire lifetime
of the app process.

A side effect of this is that app dir can now be replaced
with a temp dir in tests.

Also added a test for AppDir but it has any value
only on linux-based platforms.