frontend: add support for testing the UI
Tests are run using jest and ts-jest preprocessor.
Because the app is based on preact, this commit proposes
to test components rendering and their state with
preact-render-spy instead of enzyme.

See https://www.npmjs.com/package/preact-render-spy
for more details about enzyme and this package.

To prove that the setup works, I randomly picked
a couple simple modules and wrote a few tests for them.

See changes in README.md for more details overall.