coins: use big.Int for amounts
Another step to remove btc specific types from the account/coin
interfaces, to pave the way for other coins.

Only the interface types are changed. The implementation internally
still uses btcutil.Amount/int64 for the amounts, and converts only at
the interface.

The remaining instances of btcutil.Amount in the interface are in the
TxInfo in Transactions(). Those will be handled in a later PR, as the
tx list interface needs more refactoring anyway.