MAKE SOME BAS(E)S NOISE
This commits implements noise encryption and authentication with the
base. The 3 part noise 'XX' handshake is done over websocket. The
keypair of the app and the public static key of the base are saved to
a file. If the public static key of either party does not exist in
either party's configuration, the user is asked to verify the channel
binding hash on the base by comparing its value with the app. For this
an event notification is emitted, that allows the frontend to display
the channel hash data.