bitboxbase: Handle Disconnects of the RPC Connection
The backend now passes a callback to properly unregister through the
manager and bitboxbase package to the rpcclient. In case the websocket
connection fails, the callback is called and both read and write
websocket loops are closed.

Also fixes the ordering of the close channel to make sure that no double
closing, or writing into a closed websocket connection occurs.