Use GOB RPC to communicate with the BitBox Base
Implement a RPC client that is capable of communicating with the
BitBox Base RPC server. The rpc connection is intercepted with an
io.ReadWriteCloser that is capable of receiving events emitted by a
websocket send and receive channel event loop. Messages that are sent
and received are noise encrypted in this event loop. Every received
message has an opcode appended to it to identify either an rpc method's
response or a notification.

Files and code is moved from what was previously in the updater
directory to rpcclient.