qt: allow Safello iframe
QtWebEngine has very strange behavior with intercepting requests:

Clicking a link, left click on target=_blank or middle click on
anything, invokes createWindow(). Creating a page therein with a new
profile and a new request interceptor only works for middle
clicks. For left clicks, the original page's interceptor is called.

This prevents us from cleanly separating the main app from external
urls in terms of what kind of intercepting should happen, be it to
open a new Qt browser window or to launch externally.

As a workaround, we use the same interceptor for everything and take
the heuristic that there is no firstParty url for external links to
act.