base-backend: refactor initial connection and noise pairing
Because:
Previously, the '/connectbase' endpoint would not succeed until after
the noise pairing hash was confirmed by the user. This means that when
connecting to the Base, the user would not be routed to the setup
wizard until after the pairing had succeeded (which happens
automatically now, but in final implementation would mean the user
would never be routed to the wizard)

This commit:
Refactors the intial rpc client ping and websocket dial to a separate
function which succeeds on initial connection. Only after this is the
noise pairing initialized.