backend/btc: do not process callbacks on a closed account
Adter an account and database are closed, it still receives
notifications it subscribed to. It also receives outstanding responses
from queries made to the Electrum server.

The notifications should be unsubscribed and the queries cancelled
upon close, but as a quick measure, it is sufficient to not process
the callbacks after the account has closed.