frontends/qt: enable HiDPI auto-scaling on Windows
This makes the app scale well with the rest of the display
when the scaling factor is greater than 100%, especially font sizes.

Only Windows because there has been unexpected behavior on other
platforms in the past, at least on Qt versions before 5.14.
But we are currently stuck at 5.11 due to various dependencies and
older platforms support. Will reconsider and possibly upgrade to 5.14
in about a year.

See more about Qt HiDPI here:
https://doc.qt.io/qt-5/highdpi.html