backend: add UsingMobileData() to Environment
We want warn the user if a lot of headers (a few hundered megabytes)
would be downloaded over mobile data, which would be bad for users
with a limited or costly data plan. Or let the user explicitly opt in
or opt out of using the mobile data for header download.