frontend/web: fixes flickering when rotatiting through receive addresses
When the address changes there is a quick moment before the new image
is loaded and the webview does not know the new dimensions.
Altouhgh the dimension did not change, the whole content moves quickly
up and then back again when the QR code image is rendered.

Adding a min-height will force the webview to always reserve the space.