backend/keystore: don't pass full signing config to CanVerifyAddress
Only the multisig flag is needed (and only in the bitbox01).

Multisig in the app is a simple demo of p2sh-multisig with multiple
bitbox01, so it is technical debt in a sense.

When combining accounts, it is currently simpler to assume that a
device can verify the addresses of all supported types (see
SupportsAccount()), which is true for all supported devices and
account types. Multisig is the only exception, where the bitbox01 can
sign it, but the app does not implement verifying the addresses on the
2fa app.