backend/btc/maketx: extend estimateTxSize to allow mixed inputs
estimateTxSize assumed that all inputs in a tx are of the same
type (signing configuration in BitBoxApp terminology), e.g. version 0
native segwit.

It is extended to correctly handle transactions with mixed input
types. This will be needed when combining the different account
subtypes (e.g. Bitcoin Legacy, Segwit, Native Segwit) into one
account.

This is done by providing the input configuration for each input.

Testing this uncovered a bug:

signatureSize = 73 was wrong, signatureSize = 72 is correct (see
comments in the code). The unit test didn't catch it as it had too few
inputs, so by rounding the tx weight, the result was the same.

When mixing non segwit and segwit inputs, there is an empty witness
that has to be accounted for for each non-segwit input.

Tx construction, which uses this function to estimate the fees, is
still accepting only one configuration for all inputs, so it is just
repeated for now. In the next commit, mixed inputs will be allowed,
and the correct configuration per input will be used.