frontend/settings: move legacy, bech32, segwit guide entries
Since unified accounts. there are no settings mentioning those words.

We ove it to the accounts guide, but only in split accounts, showing
the relevant part.

We might delete the entries soon altogether, as there is also the
accounts description banner that says more or less the same. Keeping
it for now though to not lose all the translations, in case we need
it.

Litecoin currently does not show them, as the text has 'Bitcoin' all
over. Not gonna fix this either.