frontend/decorators/subscribe: fix lingering state of wrong endpoint
`subscribeWithoutLoading`, if false, does not get the initial by
calling the endpoint. This leads to a bug: if the component changes,
e.g. account switched, the state of the previous account lingers, even
though the subscription endpoint changed.

This fixes it by resetting the var in this case to the initial
value.

Same if `renderOnlyOnceLoaded` is false, in which case the old value
lingers until the endpoint was called. This is less dramatic as this
time is very short, but a bug nontheless.