frontend: use native-locale but prefer userLanguage when set
The frontend now picks the UI display languge in this order, whichever
is available first:

- userLanguage from the frontend app config
- native-locale from the backend API handlers
- static defaultUserLanguage

When i18next module triggers languageChanged event, the app persists
the selection in frontend.userLanguage config. It sets the userLanguage
field to the newly selected language unless it matches native-locale.
In the latter case, userLanguage is set to null to make the app use
native-language. If the platform on which the app is running changes
locale later on, the app UI language will follow.

A locale may contain region and other sub tags, like 'de-DE'.
When matching against translations known to the app in the language
selector component or languageChanged event handler, we also now do a
relaxed matching, comparing the main language tag. For example, native
locale 'pt_BR' matches 'pt' translation language.