frontend: fix focus when editing long notes
The edit mode did not always focus the input field correctly.
This could happen due to the input not be available during
re-rendering or the cursor not be in the visible area of the
text input when focusing. This change attemps a focus before
updating the state and also tries to scroll to the end of the
text input, so that the cursor is visible when attempting the
focus.